+++ projects/query-search/src/lib/components/query-search-group/query-search-group.component.ts
@@ -1,9 +1,11 @@
-import {Component, EventEmitter, HostBinding, Input, OnInit, Output, QueryList, ViewChildren} from '@angular/core';
-import {QueryGroup} from "../../models";
-import {isEven} from "../../query-search.helpers";
+import {Component, EventEmitter, HostBinding, Input, Output, QueryList, ViewChildren} from '@angular/core';
+import {QueryGroup, SavedFilter} from "../../models";
+import {isEven} from "../../helpers/query-search.helpers";
 import {QuerySearchService} from "../../query-search.service";
 import {inOutAnimations} from "../../animations";
 import {QuerySearchItemComponent} from "../query-search-item/query-search-item.component";
+import {CdkDragDrop, moveItemInArray} from "@angular/cdk/drag-drop";
+import {NameDialogComponent} from "../dialogs/name-dialog/name-dialog.component";
 
 @Component({
   selector: 'query-search-group',
@@ -13,7 +15,7 @@ import {QuerySearchItemComponent} from "../query-search-item/query-search-item.c
     ...inOutAnimations
   ]
 })
-export class QuerySearchGroupComponent implements OnInit {
+export class QuerySearchGroupComponent {
 
   @Output()
   generateClicked = new EventEmitter<boolean>(true);
@@ -26,9 +28,10 @@ export class QuerySearchGroupComponent implements OnInit {
     }
   }
 
-
   @ViewChildren(QuerySearchItemComponent) items: QueryList<QuerySearchItemComponent>;
   @ViewChildren(QuerySearchGroupComponent) children: QueryList<QuerySearchGroupComponent>;
+
+  loadingFilters = new EventEmitter<boolean>(true);
 
   get group() {
     return this._group;
@@ -56,17 +59,12 @@ export class QuerySearchGroupComponent implements OnInit {
   _showDivider = false;
   private _group: QueryGroup;
 
-  constructor(private querySearchService: QuerySearchService) {
+  constructor(public querySearchService: QuerySearchService) {
   }
-
-  ngOnInit(): void {
-    console.log(this.group);
-  }
-
 
   addItem() {
-   this.group.addItem();
-   this.querySearchService.log('Adding item to', this);
+    this.group.addItem();
+    this.querySearchService.log('Adding item to', this);
   }
 
   addGroup() {
@@ -91,6 +89,64 @@ export class QuerySearchGroupComponent implements OnInit {
     this.removed.emit(this.group.id);
   }
 
+  clear() {
+    this.querySearchService.log('Clearing group', this.group.id, this);
+    this.children.forEach(child => child.remove());
+    this.items.forEach(item => item.remove());
+    this.group.addItem();
+  }
+
+  drop(event: CdkDragDrop<string[]>) {
+    moveItemInArray(this.group.items, event.previousIndex, event.currentIndex);
+  }
+
+  loadFilter(filterName: string) {
+    this.loadingFilters.emit(true);
+
+    const filter = this.getFilter(filterName);
+
+    if (!!filter) {
+      this.children.forEach(child => child.remove());
+      this.items.forEach(item => item.remove());
+      this.group.apply(filter.ruleGroup);
+
+      this.loadingFilters.emit(false);
+    } else {
+      this.loadingFilters.emit(false);
+    }
+  }
+
+  renameFilter(filterName: string) {
+    const filter = this.getFilter(filterName);
+
+    if (!!filter) {
+      this.querySearchService.openFilterNameDialog(NameDialogComponent, filter, 'EDIT').subscribe();
+    }
+  }
+
+  deleteFilter(filterName: string) {
+    const filter = this.getFilter(filterName);
+    if (!!filter) {
+      this.querySearchService.deleteFilter(filter);
+    }
+  }
+
+  createFilter() {
+    const queryRuleGroup = this.group.filterValue;
+
+    if (Object.keys(queryRuleGroup).length === 0) {
+      this.querySearchService.log('Refusing to make filter from empty QueryRuleGroup', queryRuleGroup);
+      return;
+    }
+
+    const savedFilter: SavedFilter = {
+      name: '',
+      ruleGroup: queryRuleGroup
+    }
+
+    this.querySearchService.openFilterNameDialog(NameDialogComponent, savedFilter, 'CREATE').subscribe();
+  }
+
 
   get toolbarColor() {
     if (!!this._group) {
@@ -102,6 +158,18 @@ export class QuerySearchGroupComponent implements OnInit {
     }
 
     return 'primary';
+  }
+
+  get groupColor() {
+    if (!!this._group) {
+      if (this._group.depth > 0 && isEven(this._group.depth)) {
+        return 'primary';
+      } else if (this._group.depth > 0) {
+        return 'warn';
+      }
+    }
+
+    return 'accent';
   }
 
   get disableTypeButtons() {
@@ -116,4 +184,15 @@ export class QuerySearchGroupComponent implements OnInit {
     return this.group.depth === 0;
   }
 
+  private getFilter(filterName: string) {
+    const filter = this.querySearchService.savedFilters.value.find(f => f.name === filterName);
+
+    if (!!filter) {
+      return filter
+    }
+
+    this.querySearchService.log('Could not find filter for name', filterName);
+
+    return null;
+  }
 }