+++ projects/query-search/src/lib/models/index.ts
@@ -5,3 +5,4 @@ export * from './query-group.model';
 export * from './rules';
 export * from './provided-value.model';
 export * from './value-notification.model';
+export * from './saved-filter.model';