+++ projects/query-search/src/lib/models/name-dialog-data.model.ts
@@ -1,0 +1,6 @@
+import {SavedFilter} from "./saved-filter.model";
+
+export interface NameDialogData {
+  action: 'EDIT' | 'CREATE';
+  filter: SavedFilter;
+}