+++ projects/query-search/src/lib/models/saved-filter.model.ts
@@ -1,0 +1,6 @@
+import {QueryRuleGroup} from "./rules";
+
+export interface SavedFilter {
+  name: string;
+  ruleGroup: QueryRuleGroup
+}