+++ projects/query-search/src/lib/query-search.service.ts
@@ -1,21 +1,30 @@
-import {EventEmitter, Inject, Optional} from '@angular/core';
+import {EventEmitter, Inject, Optional, TemplateRef} from '@angular/core';
 import {Injectable} from '@angular/core';
-import {QueryField, ValueNotification, QueryRuleGroup} from "./models";
-import {BehaviorSubject} from "rxjs";
+import {QueryField, ValueNotification, QueryRuleGroup, SavedFilter} from "./models";
+import {BehaviorSubject, Observable} from "rxjs";
 import {QUERY_SEARCH_CONFIG, QuerySearchConfiguration} from "./query-search.config";
 import {ConditionOperator} from "./enums";
+import {NameDialogData} from "./models/name-dialog-data.model";
+import {MatDialog} from "@angular/material/dialog";
+import {map} from "rxjs/operators";
+import {ComponentType} from "@angular/cdk/overlay";
 
 @Injectable({
   providedIn: 'root'
 })
 export class QuerySearchService {
 
+  savedFilters = new BehaviorSubject<SavedFilter[]>([]);
   queryUpdated = new EventEmitter<any>();
   operators: string[] = [];
   fields: BehaviorSubject<QueryField[]> = new BehaviorSubject<QueryField[]>([]);
   valueFieldDidChange: EventEmitter<ValueNotification> = new EventEmitter<ValueNotification>(true);
   searchValueFieldDidChange: EventEmitter<ValueNotification> = new EventEmitter<ValueNotification>(true);
+  savedFilterUpdated: EventEmitter<SavedFilter> = new EventEmitter<SavedFilter>(true);
+  savedFilterCreated: EventEmitter<SavedFilter> = new EventEmitter<SavedFilter>(true);
+  savedFilterDeleted: EventEmitter<SavedFilter> = new EventEmitter<SavedFilter>(true);
 
+  private _hasSavedFilters = false;
   private _fields: QueryField[] = [];
   private readonly _loggingCallback: (...args: any[]) => void = () => {
   };
@@ -29,7 +38,8 @@ export class QuerySearchService {
   private readonly _showFieldNameSuffix: boolean = true;
   private readonly _showOperatorSuffix: boolean = true;
 
-  constructor(@Optional() @Inject(QUERY_SEARCH_CONFIG) configuration: QuerySearchConfiguration) {
+  constructor(@Optional() @Inject(QUERY_SEARCH_CONFIG) configuration: QuerySearchConfiguration,
+              private matDialog: MatDialog) {
     this.operators = Object.keys(ConditionOperator).filter(k => !k.includes('LOW'));
     this._loggingCallback = configuration.loggingCallback;
     this._debug = configuration.debug;
@@ -127,6 +137,65 @@ export class QuerySearchService {
     this.addFields(fields);
   }
 
+  /**
+   * Provide saved filters to use for later
+   * @param filters
+   */
+  provideSavedFilters(filters: SavedFilter[] | Observable<SavedFilter[]>) {
+    if (filters instanceof Observable) {
+      filters.subscribe(this.savedFilters);
+    } else {
+      this.savedFilters.next(filters);
+    }
+
+    this._hasSavedFilters = true;
+  }
+
+  /**
+   * Edit a saved filter
+   * @param componentOrTemplateRef
+   * @param filter
+   * @param action
+   */
+  openFilterNameDialog<T>(componentOrTemplateRef: ComponentType<T> | TemplateRef<T>, filter: SavedFilter, action: 'EDIT' | 'CREATE') {
+    const previousFilterName = filter.name;
+    const data: NameDialogData = {
+      filter,
+      action
+    };
+
+    return this.matDialog.open(componentOrTemplateRef, {
+      data,
+      height: '225px',
+      width: '400px',
+    }).afterClosed().pipe(
+      map(result => {
+        if (action === 'EDIT' && !!result && previousFilterName !== filter.name) {
+          this.log('Saved filter updated', filter);
+          this.savedFilterUpdated.emit(filter);
+          return filter;
+        } else if (action === 'CREATE' && !!result && previousFilterName !== filter.name) {
+          this.log('New filter saved', filter);
+          this.savedFilters.next([filter, ...this.savedFilters.value]);
+          this.savedFilterCreated.emit(filter);
+          return filter;
+        }
+
+        return null;
+      })
+    )
+  }
+
+  /**
+   * Delete a saved filter
+   * @param filter
+   */
+  deleteFilter(filter: SavedFilter) {
+    this.log('Filter deleted', filter);
+    this.savedFilters.next(this.savedFilters.value.filter(f => f.name !== filter.name));
+    this.savedFilterDeleted.emit(filter);
+  }
+
   get generateButtonText(): string {
     return this._generateButtonText;
   }
@@ -149,5 +218,9 @@ export class QuerySearchService {
 
   get showOperatorSuffix() {
     return this._showOperatorSuffix;
+  }
+
+  get hasSavedFilters() {
+    return this._hasSavedFilters;
   }
 }